# specs/004-ai-chat-todo/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Phase 3 - AI Chat-Driven Todo API
  version: 1.2.0
  description: |
    Backend API for the AI Chat-Driven Todo Application.
    Authentication is handled via standard JWT endpoints.
    The FastAPI backend verifies JWT tokens and exposes task CRUD,
    a ChatKit-compatible chat endpoint, and calendar data endpoints.
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            required:
              - loc
              - msg
              - type
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Location of the validation error
                example: ["body", "email"]
              msg:
                type: string
                description: Validation error message
                example: "value is not a valid email address"
              type:
                type: string
                description: Type of validation error
                example: "value_error.email"
    Priority:
      type: string
      enum: [high, medium, low]
      default: medium
    TaskStatus:
      type: string
      enum: [pending, completed]
      default: pending
    TagResponse:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    TaskCreate:
      type: object
      required: [title, due_date]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        due_date:
          type: string
          format: date
        priority:
          $ref: '#/components/schemas/Priority'
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        due_date:
          type: string
          format: date
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
    TaskResponse:
      type: object
      required: [id, user_id, title, due_date, priority, status, created_at, tags]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        due_date:
          type: string
          format: date
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskListResponse:
      type: object
      required: [tasks, total]
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        total:
          type: integer
...
    UserProfileResponse:
      type: object
      required: [id, email, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
paths:
  /api/auth/register:
    post:
      summary: Create a new user account
      description: Register a new user with email and password, returning JWT token and user details
      tags:
        - Authentication
      security: []  # No authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: user@example.com
                password:
                  type: string
                  minLength: 1
                  description: User's password (minimum 1 character)
                  example: securepassword123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - token
                  - user
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/UserProfileResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/auth/login:
    post:
      summary: Authenticate existing user
      description: Login with email and password, returning JWT token and user details
      tags:
        - Authentication
      security: []  # No authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: user@example.com
                password:
                  type: string
                  description: User's password
                  example: securepassword123
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - token
                  - user
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/auth/me:
    get:
      summary: Get current user profile
      description: Retrieve the profile of the currently authenticated user
      tags:
        - Authentication
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/me:
    get:
      summary: Get current user profile
      tags: [Auth]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/tasks:
    get:
      summary: List tasks with optional filters
      tags: [Tasks]
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: tag_ids
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
          style: form
          explode: true
        - name: search
          in: query
          description: Keyword search across title, description, tags
          schema:
            type: string
        - name: due_date_from
          in: query
          schema:
            type: string
            format: date
        - name: due_date_to
          in: query
          schema:
            type: string
            format: date
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [due_date, priority, title]
            default: due_date
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Filtered task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Unauthorized

    post:
      summary: Create a new task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /api/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get a single task
      tags: [Tasks]
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized
        '404':
          description: Task not found

    patch:
      summary: Update a task
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized
        '404':
          description: Task not found
        '422':
          description: Validation error

    delete:
      summary: Delete a task
      tags: [Tasks]
      responses:
        '204':
          description: Task deleted
        '401':
          description: Unauthorized
        '404':
          description: Task not found
...